requirements.txt
fastapi
uvicorn
yt_dlp
faster-whisper
google-generativeai
python-multipart